+++ api/grpc_server.go
@@ -34,7 +34,7 @@ func (s SpaceMeshGrpcService) StopService() {
 func NewGrpcService() *SpaceMeshGrpcService {
 	port := config.ConfigValues.GrpcServerPort
 	server := grpc.NewServer()
-	return &SpaceMeshGrpcService{Server: server, Port: port}
+	return &SpaceMeshGrpcService{Server: server, Port: uint(port)}
 }
 
 func (s SpaceMeshGrpcService) StartService(started chan bool) {
